refactor: move to test to separate file.
